requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
openai~=1.82.1

requests~=2.32.3
pydantic~=2.11.5
protobuf~=6.30.2
torch~=2.7.1
torchvision~=0.22.1
numpy~=2.3.1
pillow~=11.2.1